transfer to school mac

+++ cpp_module_09/main.cpp
@@ -6,7 +6,7 @@
 #include <algorithm>
 #include <limits>
 
-const size_t jacobsthalGroupSize[] = {
+const unsigned int jacobsthalGroupSize[] = {
 	2, 2, 6, 10, 22, 42, 86, 170, 342, 682, 1366,
         2730, 5462, 10922, 21846, 43690, 87382, 174762, 349526, 699050,
         1398102, 2796202, 5592406, 11184810, 22369622, 44739242, 89478486,
